panel/requirements.txt
Flask==3.0.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.0.1
python-dateutil==2.8.2
